GTS-BE/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
idna==3.3
mysqlclient==2.1.0
pipreqs==0.4.11
pycodestyle==2.8.0
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
yarg==0.1.9
